requirements/production_server.txt

# For smarter timezone handling
pytz==2013.9

# django packages
Django==1.7
django-timezone-field==1.1
djangorestframework==2.4.4
django-extensions==1.3.11
django-filter==0.9.1
django-rest-swagger==0.2.7

# Sphinx documentation
Sphinx==1.2
sphinx-rtd-theme==0.1.6

# mptt
django-mptt==0.6.1
django-mptt-admin==0.1.10

# django suit
django-suit==0.2.10
		
# postgres adapter for python
psycopg2==2.5.3

#uwsgi
uwsgi==2.0.9

